Parts 1 and 2 of 7 pass.

+++ Days/Day7/Day07Tests/PositionCost.cs
@@ -5,9 +5,9 @@ namespace Day07Tests
     internal class PositionCost
     {
         public int Position { get; set; }
-        public int Fuel { get; set; }
+        public Int64 Fuel { get; set; }
 
-        public PositionCost(int position, Int32 cost = Int32.MaxValue)
+        public PositionCost(int position, Int64 cost = Int32.MaxValue)
         {
             this.Position = position;
             this.Fuel = cost;
